Adding EnterDungeonPM
This is a screen in between clicking on a level and loading the level.
It might show thigns like the level difficulty, monsters or rewards
inside the dungeon, etc. It's also where the player will choose
expedition units to take into the dungeon. Right now there is just a
button to actually load the dungeon.