Added animated version of Heart tile
Added animated version of Heart tile for testing and discussions for how to wire animated symbols up.
I have changed the Canvas mode to "Screen Space - Camera", I can't see the particles if it's set to Overlay (they always render behind the whole canvas).
I added a Canvas to the Icon to get the Icon to override the sorting - I'm not aware of a better way to get the icon topmost (yet).
The animation has 2 layers, so I can animate the rays at a different rate to the icon.
Most of this is proof of concept, and give you a more heavy symbol to use for testing.
I advise we disable the OnState game object, because the particles are nested, and they won't be affected by the Canvas Group alpha, and would carry on simulating (but I could try controlling by the animator).
Note I have not done any optimization there are some easy improvments I can make for that.