Adjusted Combat scene layout to the reference resolution
Changed the combat scene to 1080x1920 as the refrence resolution. Adjusted the layout to work better with that. Adjusted the Enemy view so it works better at 16:9 and 4:3.
Fixed weird NAN issue I was having with the DungeonRewardView in the Combat scene.
Adjusted the animations for the wave transition to not have a pause when pressing button (and made animations snappier).
Adjusted the tiles (static and animated) to be full size to the artwork.
Improved layout of the Main Menu so player info doesn't overlap the buttons.