Added images for the ON states for the animated versions of the tiles
In the GamePieceView prefab I adjusted the art and positions for each symbol so they look correct.
For the OFF state I removed the icon GameObject and used a Flat image with the Background and Icon - The reason for this is each tile had a shadow, and by trying to make layered versions with different art, I might as well use a 'baked' version.
Ideally, when the tile is ON, the OFF would be disabled.
I need to fix the animations because they're currently squishing the icons