Made tile animation generic and Fixed issue with Animator state
I changed how the scaling on the animation for the ON state is done so it can be used for all the tiles.
Unity 5.6 introduces a new behaviour for the Animator (some are calling it a but) where it remembers the state for the Animator when disabling/enabling it (previous versions it would reset). I have created a simple script based on suggestions from other users with this issue. I called it AnimatorResetter and put it in the root - Joe feel free to change/move this :D