Combo Dropped animation on tiles
I have added an animation for when the combo gets dropped. I haven't tuned it yet, will need to make adjustments once I see it in game. Will see if I can figure out how to wire it up in the code. The animation is triggerd from the On controller "ComboDropped" is the trigger name.