Wired up the ChainDropped and ChainComplete tile animations
Joe, I've had a go at wiring up the animations in the code. Feel free to change or improve how I've done it. There is a bug with how it's currently set up because you can keep dragging tiles before the animation is complete, and it replays them all - you can select all the tiles doing this (need to check if the tile was one of the selected ones before playing the animation, or something like that).