Improved how the animations are triggered
Now the chain_complete and chain_dropped event is only sent to the pieces in the mChain list. This prevents all the previously triggered (but not complete) animations from restarting.
There is another issue where all the animations stop when the first one does - this interrupts any animations that started later. Maybe the event triggered by the Animation needs to specify the game tile some how...